+++ examples/router.py
@@ -9,7 +9,7 @@ from tortoise.models import Model
 
 
 class Event(Model):
-    id = fields.IntField(pk=True)
+    id = fields.IntField(primary_key=True)
     name = fields.TextField()
     datetime = fields.DatetimeField(null=True)